create config.py and function to find 7zip
